move client hooks to components
