+++ src/app/(site)/about/page.js
@@ -1,42 +1,21 @@
-"use client";
 import Image from "next/image";
-import {useEffect, useState} from "react";
 import {CreditCard, Lock, ShieldUser} from "lucide-react";
-import animateOnObserve from "@/lib/animateOnObserve";
 import Commitments from "@/components/Commitments";
 import CoreValues from "@/components/CoreValues";
 import AboutHeroSection from "@/components/AboutHeroSection";
 import TeamSection from "@/components/TeamSection";
+import AboutPageAnimations from "@/components/AboutPageAnimations";
 import {client} from "@/sanity/lib/client";
 import {aboutPageQuery} from "@/sanity/lib/queries";
 
-export default function AboutPage() {
-    const [data, setData] = useState(null);
-
-    useEffect(() => {
-        client.fetch(aboutPageQuery).then(setData);
-    }, []);
-
-    useEffect(() => {
-        // Set up animations after DOM is ready
-        const swingObserver = animateOnObserve('.swing-in-top-fwd-2');
-        const borderObserver = animateOnObserve('.border-draw');
-        const puffObserver = animateOnObserve('.puff-in-center');
-        const slideObserver = animateOnObserve('.slide-in-bottom');
-
-        // Cleanup function to disconnect observers
-        return () => {
-            swingObserver.disconnect();
-            puffObserver.disconnect();
-            borderObserver.disconnect();
-            slideObserver.disconnect();
-        };
-    }, []);
+export default async function AboutPage() {
+    const data = await client.fetch(aboutPageQuery);
 
     return (
         <div className="w-full bg-[#373737]">
-            <AboutHeroSection data={data?.heroSection} />
-            <TeamSection data={data?.teamSection} />
+            <AboutPageAnimations/>
+            <AboutHeroSection data={data?.heroSection}/>
+            <TeamSection data={data?.teamSection}/>
             <div>
                 {/*Office pictures*/}
                 <section className="pt-12">
