+++ src/app/(site)/completed-projects/page.js
@@ -1,10 +1,6 @@
-"use client";
-
 import Projects from "@/components/Projects";
 
 export default function CompletedProject() {
-
-  return (
-    <Projects pageTitle="Thi công thực tế"/>
-  );
+  return <Projects pageTitle="Thi công thực tế"/>;
 }
+
