+++ src/components/ui/banner.js
@@ -1,3 +1,5 @@
+"use client";
+
 import { debounce } from "lodash";
 import { useCallback, useEffect, useRef, useState } from "react";
 import Image from 'next/image';
